fix(frontend): [SCRUM-37-2] restructure component, add index export
