--- src/app/components/index.tsx
@@ -1,1 +1,0 @@
-export { default as Button } from './button';
