+++ src/components/index.tsx
@@ -1,0 +1,4 @@
+export { default as Button } from './button';
+export { default as Banner } from './Banner';
+export { default as NavBar } from './NavBar';
+export { default as UserProfileCard } from './UserLabel';
